Select atoms for features on 1 or >2 atoms
If a feature acts on 2 atoms, create a pseudobond
and select it when the feature is selected. Otherwise,
store the set of atoms the feature acts on and select
those when the feature is selected (unless the feature
has children that act on 2 atoms). This allows us to
do something useful with EM or bounding box restraints,
for example.