Initial session support for ref'd Chimera objects
Feature and hierarchy nodes reference Chimera bonds
or atoms so that they can be selected. We can't save
these directly in sessions as that would introduce
a circular dependency (each references the model)
so instead store the type, model ID and indices,
and regenerate the Chimera objects as needed after
session restore.