Use RMF C++ code to read trajectories
Use a new function in RMF itself to read XYZ
coordinates for trajectories. This reads into
a NumPy array which can then be passed directly
to ChimeraX. This makes the 'rmf readtraj' command
much faster, as there is no need to switch between
C++ and Python for every node as we traverse the
tree in the RMF file.